kd_pipeline/requirements.txt
# Qwen3-VL KD pipeline — pin versions in team CI as needed
torch>=2.1.0
transformers>=4.57.0
accelerate>=0.30.0
peft>=0.11.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.66.0
safetensors>=0.4.0
